Added score for incidents around ccp -> dest line
