+++ .gitignore
@@ -8,3 +8,4 @@ Incident_ranking_function.zip
 
 output/*
 !output/_dev_tests
+output/_dev_tests/line*
